Add ability to shoot a bullet

+++ src/com/chalmers/ZombieKillah/Bullet.java
@@ -5,8 +5,8 @@ package com.chalmers.ZombieKillah;
  */
 public class Bullet extends MovableObject {
 
-    public Bullet(Direction direction,int x, int y){
-        super("Bullet",x,y);
+    public Bullet(Direction direction, double x, double y){
+        super("Bullet", x, y);
         this.speed = 0.5;
         this.direction = direction;
     }
